Pathfinding is now responsability of the WalkableObject
