--- Runtime/World/Walkable.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 16349239f6b65f645913bcdeb3b00750
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
